+++ google-oauth/functions/google/getAccessToken.ts
@@ -1,0 +1,31 @@
+import type { GoogleAccessTokenJson, GoogleCredentialsJson } from "../../types.ts";
+
+export default async function getAccessToken(
+	googleJson: GoogleCredentialsJson,
+	code: string,
+): Promise<GoogleAccessTokenJson> {
+	const url = "https://oauth2.googleapis.com/token";
+	const init: RequestInit = {
+		method: "POST",
+		headers: {
+			"Content-Type": "application/x-www-form-urlencoded",
+		},
+		body: new URLSearchParams({
+			"client_id": googleJson.web.client_id,
+			"client_secret": googleJson.web.client_secret,
+			"redirect_uri": googleJson.web.redirect_uris[0],
+			code,
+			"grant_type": "authorization_code",
+		}),
+	};
+	const response = await fetch(url, init);
+	if (response.ok) {
+		const json = await response.json() as GoogleAccessTokenJson;
+		if (typeof json.refresh_token === "undefined") {
+			throw new Error("Unexpected missing Refresh Token");
+		}
+		return json;
+	} else {
+		throw new Error(`Error getting Access Token - ${response.status}`);
+	}
+}
